update home.js and dataTables.js
feat: Add tableComplete event listener

This commit adds a tableComplete event listener to the home.js file. When the tableComplete event is triggered, the doSomething function is called. This allows for additional functionality to be executed after the table has completed loading.

Changes import for setupDataTable function from @scripts/dataTables
- Added tableComplete event listener to document object
- Added doSomething function to handle tableComplete event
- Updated setupDataTable function to return the DataTable instance